requirements.txt
openpyxl
pillow
pypdf
python-magic
python-ffmpeg
python-docx
